pyscript/requirements.txt
ruamel.yaml==0.17.21
sqlalchemy==2.0.7
pytz
HASS-data-detective
pandas
numpy